From: "Mark Hammond" <MHammond@skippinet.com.au> Date: Fri, 4 Oct 1996 09:08:19 +1000
A couple of things.  As I mentioned a while back, I have made the
changes to the registry support, in getpath_nt.c.  To recap, there can be:
...\pythonpath = default core Pythonpath
...\pythonpath\Pythonwin = c:\somewhere
etc.

The code simply appends them all.  The order can not be guaranteed
(registry limitation) but the "default" is always at the end.

The main reasons for change were the length of the path, but mainly
so an uninstaller can do the right thing.