#10197: Update get[status]output versionchanged with actual version.
This was perhaps on the border between a bug fix and a feature since
the Python3 docs did not originally say it was unix only.  However, the
functions never worked anywhere but unix, and the docs were changed to
say it was unix only well before the windows support was added.

Unfortunately, windows support was added in 3.3.4 as well as 3.4.
That leaves us in the uncomfortable position of needing the tag to say
"version changed: 3.3.4" :(